requirements.txt
Django==1.6.2
argparse==1.2.1
wsgiref==0.1.2
South==0.8.4
django-tinymce==1.5.2
